+++ java_starter_package/src/main/java/Scorer.java
@@ -1,10 +1,8 @@
-import java.util.List;
-
 /**
  * Created by Silvestrov Ilya
  * Date: Oct 9, 2010
  * Time: 12:00:51 AM
  */
 public interface Scorer {
-    int score(PlanetWarsState state, Plan plan);
+    long score(PlanetWarsState state, Plan plan);
 }
